+++ frontend/js/components/input_controls.ts
@@ -82,6 +82,9 @@ template.innerHTML = String.raw`
       <button title="Open info menu" id="info-button" class="button">
         <span class="fas ${ICONS.info}"></span>
       </button>
+      <button title="Open help menu" id="help-button" class="button">
+        <span class="fas ${ICONS.help}"></span>
+      </button>
       <button title="Open settings menu" id="settings-button" class='button'>
         <span class="fas ${ICONS.settings}"></span>
       </button>
@@ -100,6 +103,7 @@ export class InputControls extends HTMLElement {
   private toggleMarkerButton: HTMLButtonElement;
   private chromosomeViewButton: HTMLButtonElement;
   private infoButton: HTMLButtonElement;
+  private helpButton: HTMLButtonElement;
   private settingsButton: HTMLButtonElement;
 
   private searchButton: HTMLButtonElement;
@@ -111,6 +115,7 @@ export class InputControls extends HTMLElement {
   private onToggleMarker: () => void;
   private onOpenSettings: () => void;
   private onOpenInfo: () => void;
+  private onOpenHelp: () => void;
   private onToggleChromView: () => void;
   private onSearch: (query: string) => Promise<ApiSearchResult | null>;
   private onChange: (settings: RenderSettings) => void;
@@ -122,6 +127,7 @@ export class InputControls extends HTMLElement {
     onPositionChange: (newXRange: [number, number]) => void,
     onOpenSettings: () => void,
     onOpenInfo: () => void,
+    onOpenHelp: () => void,
     onToggleChromView: () => void,
     onSearch: (query: string) => Promise<ApiSearchResult | null>,
     onChange: (settings: RenderSettings) => void,
@@ -129,6 +135,7 @@ export class InputControls extends HTMLElement {
     this.session = session;
     this.onOpenSettings = onOpenSettings;
     this.onOpenInfo = onOpenInfo;
+    this.onOpenHelp = onOpenHelp;
     this.onToggleChromView = onToggleChromView;
     this.getMarkerOn = () => this.session.getMarkerModeOn();
     this.onToggleMarker = () => this.session.toggleMarkerMode();
@@ -177,6 +184,7 @@ export class InputControls extends HTMLElement {
 
     this.chromosomeViewButton = this.querySelector("#chromosome-view-button");
     this.infoButton = this.querySelector("#info-button");
+    this.helpButton = this.querySelector("#help-button");
     this.settingsButton = this.querySelector("#settings-button");
 
     this.searchButton = this.querySelector("#search");
@@ -192,6 +200,10 @@ export class InputControls extends HTMLElement {
     this.infoButton.addEventListener("click", () => {
       this.onOpenInfo();
     });
+
+    this.helpButton.addEventListener("click", () => {
+      this.onOpenHelp();
+    })
 
     // FIXME: Also enter when inside the input?
     this.searchButton.addEventListener("click", () => {